Internal

Internal

Constraints.all_different — Constant

_all_different

Global constraint ensuring that all the values of a given configuration are unique.

Constraints.all_equal — Constant

all_equal

Global constraint ensuring that all the values of x are all equal.

Constraints.all_equal_param — Constant

all_equal_param

Global constraint ensuring that all the values of x are all equal to a given parameter param.

Constraints.dist_different — Constant

dist_different

Local constraint ensuring that concept(dist_different, x) = |x[1] - x[2]| ≠ |x[3] - x[4]|).

Constraints.eq — Constant

eq

Equality between two variables.

Constraints.less_than_param — Constant

less_than_param

Constraint ensuring that the value of x is less than a given parameter param.

Constraints.minus_equal_param — Constant

minus_equal_param

Constraint ensuring that the difference between x[1] and x[2] is equal to a given parameter param.

Constraints.ordered — Constant

ordered(x)

Global constraint ensuring that all the values of x are ordered.

Constraints.sequential_tasks — Constant

sum_equal_param

Constraint ensuring that the start and completion times of two tasks are not intersecting: x[1] ≤ x[2] || x[3] ≤ x[4].

Constraints.sum_equal_param — Constant

sum_equal_param

Global constraint ensuring that the sum of the values of x is equal to a given parameter param.