Public

Public

Constraints.usual_constraints — Constant

usual_constraints::Dict

Dictionary that contains all the usual constraints defined in Constraint.jl.

Constraints.usual_symmetries — Constant

symmetries

A Dictionary that contains the function to apply for each symmetry to avoid searching a whole space.

Constraints.Constraint — Type

Constraint

Parametric stucture with the following fields.

concept: a Boolean function that, given an assignment x, outputs true if x satisfies the constraint, and false otherwise.

error: a positive function that works as preferences over invalid assignements. Return 0.0 if the constraint is satisfied, and a strictly positive real otherwise.

Constraints.args_length — Method

args_length(c::Constraint)

Return the expected length restriction of the arguments in a constraint c. The value nothing indicates that any strictly positive number of value is accepted.

Constraints.concept — Method

concept(c::Constraint)

Return the concept (function) of constraint c. concept(c::Constraint, x...; param = nothing) Apply the concept of c to values x and optionally param.

Constraints.error_f — Method

error_f(c::Constraint)

Return the error function of constraint c. error_f(c::Constraint, x; param = nothing) Apply the error function of c to values x and optionally param.

Constraints.params_length — Method

params_length(c::Constraint)

Return the expected length restriction of the arguments in a constraint c. The value nothing indicates that any strictly positive number of parameters is accepted.

Constraints.symmetries — Method

symmetries(c::Constraint)

Return the list of symmetries of c.